+++ test/integration/test_agent_sink.go
@@ -1,4 +1,4 @@
-package e2e
+package integration
 
 import (
 	"encoding/json"
